+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symboltable/VariableNameDeclaration.java
@@ -36,6 +36,12 @@ public class VariableNameDeclaration extends AbstractNameDeclaration implements 
         }
     }
 
+    public boolean isVarargs() {
+        ASTVariableDeclaratorId astVariableDeclaratorId = (ASTVariableDeclaratorId) node;
+        ASTFormalParameter parameter = astVariableDeclaratorId.getFirstParentOfType(ASTFormalParameter.class);
+        return parameter != null && parameter.isVarargs();
+    }
+
     public boolean isExceptionBlockParameter() {
         return ((ASTVariableDeclaratorId) node).isExceptionBlockParameter();
     }